requirements.txt
h5py==3.1.0
numpy==1.19.2
Pillow==9.2.0
torch==1.9.0
torchvision==0.10.0
